added icon instead of word comment
